+++ index.js
@@ -1,28 +1,23 @@
 // Required SDK
 
-let { ToggleButton } = require('sdk/ui/button/toggle');
-let panels = require("sdk/panel");
+const { ToggleButton } = require('sdk/ui/button/toggle'),
+	{ Panel: NewPanel } = require("sdk/panel"),
+	{ Request } = require("sdk/request"),
+	self = require("sdk/self"),
+	sp = require('sdk/simple-prefs'),
+	clipboard = require("sdk/clipboard"),
+	tabs = require("sdk/tabs"),
+	{ browserWindows: windows} = require("sdk/windows"),
+	ContextMenu = require("sdk/context-menu"),
+	{setInterval, setTimeout, clearInterval} = require("sdk/timers"),
+	{ get: _} = require("sdk/l10n"),
+	notifications = require("sdk/notifications");
 
-let Request = require("sdk/request").Request;
-
-let self = require("sdk/self");
-let sp = require('sdk/simple-prefs');
 let simplePrefs = require('sdk/simple-prefs').prefs;
-let clipboard = require("sdk/clipboard");
 
-function check_delay_onChange(){
-	if(getPreference("check_delay") <1){
-		savePreference("check_delay") = 1;
-	}
+function Request_Get(options){
+	Request(options).get();
 }
-sp.on("check_delay", check_delay_onChange);
-
-let tabs = require("sdk/tabs");
-let windows = require("sdk/windows").browserWindows;
-
-let ContextMenu = require("sdk/context-menu");
-
-let {setInterval, setTimeout, clearInterval} = require("sdk/timers");
 
 function getPreference(prefId){
 	return simplePrefs[prefId];
@@ -33,6 +28,13 @@ function savePreference(prefId, value, updatePanel){
 		refreshPanel();
 	}
 }
+
+function check_delay_onChange(){
+	if(getPreference("check_delay") <1){
+		savePreference("check_delay") = 1;
+	}
+}
+sp.on("check_delay", check_delay_onChange);
 
 function getBooleanFromVar(string){
 	switch(typeof string){
@@ -55,28 +57,31 @@ function getBooleanFromVar(string){
 	}
 }
 
-let myIconURL = self.data.url("live_offline_64.svg");
-let myIconURL_16 = self.data.url("live_offline_16.svg");
+const myIconURL = self.data.url("live_offline_64.svg"),
+	myIconURL_16 = self.data.url("live_offline_16.svg"),
+	myIconURL_online = {
+		"16": "./live_online_16.svg",
+		"32": "./live_online_32.svg",
+		"64": "./live_online_64.svg"
+	},
+	myIconURL_offline = {
+		"16": "./live_offline_16.svg",
+		"32": "./live_offline_32.svg",
+		"64": "./live_offline_64.svg"
+	};
 
-let myIconURL_online = {
-	"16": "./live_online_16.svg",
-	"32": "./live_online_32.svg",
-	"64": "./live_online_64.svg"
-}
-
-
-let myIconURL_offline = {
-	"16": "./live_offline_16.svg",
-	"32": "./live_offline_32.svg",
-	"64": "./live_offline_64.svg"
-}
-
-let _ = require("sdk/l10n").get;
-
-let websites = ["beam","dailymotion","hitbox","twitch"];
+let websites = {};
 let liveStatus = {};
 let channelInfos = {};
-for(website of websites){
+
+(function(){
+	let websitesToLoad = ["beam","dailymotion","hitbox","twitch"];
+	for(let i in websitesToLoad){
+		websites[websitesToLoad[i]] = require(`./data/js/platforms/${websitesToLoad[i]}.js`)
+	}
+})();
+
+for(let website in websites){
 	liveStatus[website] = {};
 	channelInfos[website] = {};
 }
@@ -88,10 +93,10 @@ if(typeof getPreference("dailymotion_check_delay") == "number" && getPreference(
 	delete simplePrefs["dailymotion_check_delay"];
 }
 if(getPreference("stream_keys_list") == ""){
-	let importSreamsFromOldVersion = function(){
+	(function(){
 		let somethingElseThanSpaces = /[^\s]+/;
 		let newPrefTable = new Array();
-		for(let website of websites){
+		for(let website in websites){
 			let pref = getPreference(`${website}_keys_list`);
 			if(typeof pref != "undefined" && pref != "" && somethingElseThanSpaces.test(pref)){
 				let myTable = pref.split(",");
@@ -101,14 +106,13 @@ if(getPreference("stream_keys_list") == ""){
 			}
 		}
 		savePreference("stream_keys_list", newPrefTable.join(", "), false);
-		for(let website of websites){
+		for(let website in websites){
 			let pref = getPreference(`${website}_keys_list`);
 			if(typeof pref != "undefined"){
 				delete simplePrefs[`${website}_keys_list`];
 			}
 		}
-	}
-	importSreamsFromOldVersion();
+	})();
 }
 if(typeof getPreference("livenotifier_version") == "string" && getPreference("livenotifier_version") != "0.0.0"){
 	delete simplePrefs["livenotifier_version"];
@@ -156,8 +160,8 @@ class streamListFromSetting{
 		}
 		
 		let obj = {};
-		for(let i in websites){
-			obj[websites[i]] = {};
+		for(let website in websites){
+			obj[website] = {};
 		}
 		if(pref != "" && somethingElseThanSpaces.test(pref)){
 			let myTable = pref.split(",");
@@ -368,7 +372,6 @@ function getStreamURL(website, id, contentId, usePrefUrl){
 	}
 }
 
-//Définition du bouton icon: self.data.url("logo.svg")
 var firefox_button = ToggleButton({
 	id: "streamnotifier_button",
 	label: _("Stream offline"),
@@ -377,10 +380,10 @@ var firefox_button = ToggleButton({
 	onClick: handleChange
 });
 
-var panel = panels.Panel({
+var panel = NewPanel({
 	height: 350,
 	width: 285,
-	contentScriptFile: [self.data.url("perfect-scrollbar.min.js"), self.data.url("options-data.js"), self.data.url("panel_contentScriptFile.js")],
+	contentScriptFile: [self.data.url("js/perfect-scrollbar.min.js"), self.data.url("js/options-data.js"), self.data.url("js/panel_contentScriptFile.js")],
 	contentURL: self.data.url("panel.html"),
 });
 
@@ -396,7 +399,7 @@ function refreshStreamsFromPanel(){
 	setTimeout(waitToUpdatePanel, 5000);
 }
 
-let addStream_URLpatterns = {"dailymotion": [/^(?:http|https):\/\/games\.dailymotion\.com\/(?:live|video)\/([a-zA-Z0-9]+).*$/, /^(?:http|https):\/\/www\.dailymotion\.com\/(?:embed\/)?video\/([a-zA-Z0-9]+).*$/, /^(?:http|https):\/\/games\.dailymotion\.com\/[^\/]+\/v\/([a-zA-Z0-9]+).*$/],
+const addStream_URLpatterns = {"dailymotion": [/^(?:http|https):\/\/games\.dailymotion\.com\/(?:live|video)\/([a-zA-Z0-9]+).*$/, /^(?:http|https):\/\/www\.dailymotion\.com\/(?:embed\/)?video\/([a-zA-Z0-9]+).*$/, /^(?:http|https):\/\/games\.dailymotion\.com\/[^\/]+\/v\/([a-zA-Z0-9]+).*$/],
 		"channel::dailymotion": [/^(?:http|https):\/\/(?:games\.|www\.)dailymotion\.com\/user\/([^\s\t\/]+).*$/, /^(?:http|https):\/\/(?:games\.|www\.)dailymotion\.com\/(?!user\/|monetization\/|archived\/|fr\/|en\/|legal\/|stream|rss)([^\s\t\/]+).*$/],
 		"hitbox": [/^(?:http|https):\/\/www\.hitbox\.tv\/(?:embedchat\/)?([^\/\?\&]+).*$/],
 		"twitch": [/^(?:http|https):\/\/www\.twitch\.tv\/([^\/\?\&]+).*$/,/^(?:http|https):\/\/player\.twitch\.tv\/\?channel\=([\w\-]+).*$/],
@@ -421,7 +424,7 @@ ContextMenu.Item({
 				return false;
 			})
 	],
-	contentScriptFile: self.data.url("page_getUrlLink.js"),
+	contentScriptFile: self.data.url("js/page_getUrlLink.js"),
 	onMessage: function(data){
 		console.info(`[ContextMenu] URL: ${data}`);
 		addStreamFromPanel({"ContextMenu_URL": data});
@@ -476,10 +479,10 @@ function addStreamFromPanel(data){
 						return true;
 					} else {
 						let id_toChecked = id;
-						let current_API = new API(website, id);
+						let current_API = websites[website].API(id);
 						
 						if(website_channel_id.test(source_website)){
-							current_API = new API_channelInfos(website, `channel::${id}`);
+							current_API = websites[website].API_channelInfos(`channel::${id}`);
 						}
 						
 						Request({
@@ -489,8 +492,8 @@ function addStreamFromPanel(data){
 								let id = id_toChecked;
 								data = response.json;
 								
-								console.group()
-								console.info(`${website} - ${current_API.url}`);
+								console.group();
+								console.info(`${website} - ${response.url}`);
 								console.dir(data);
 								console.groupEnd();
 								
@@ -534,7 +537,7 @@ function addStreamFromPanel(data){
 	if(typeof data != "object" && type != "ContextMenu"){
 		if(!data.hasOwnProperty("embed_list")){
 			let page_port = current_tab.attach({
-				contentScriptFile: self.data.url("page_getEmbedList.js")
+				contentScriptFile: self.data.url("js/page_getEmbedList.js")
 			});
 			addStreamFromPanel_pageListener.push(page_port);
 			page_port.port.on("addStream", addStreamFromPanel);
@@ -688,30 +691,32 @@ function updatePanelData(){
 				}
 				
 				if(JSON.stringify(liveStatus[website][id]) == "{}"){
-					let streamData = channelInfos[website][id];
-					let contentId = id;
-					
-					//console.info(`No data found, using channel infos: ${id} (${website})`);
-					
-					let streamInfo = {
-						id: id,
-						type: "channel",
-						contentId: contentId,
-						online: streamData.online,
-						website: website,
-						streamName: streamData.streamName,
-						streamStatus: streamData.streamStatus,
-						streamGame: streamData.streamGame,
-						streamOwnerLogo: streamData.streamOwnerLogo,
-						streamCategoryLogo: streamData.streamCategoryLogo,
-						streamCurrentViewers: streamData.streamCurrentViewers,
-						streamUrl: getStreamURL(website, id, contentId, true),
-						facebookID: streamData.facebookID,
-						twitterID: streamData.twitterID
+					if(typeof channelInfos[website][id] != "undefined"){
+						let streamData = channelInfos[website][id];
+						let contentId = id;
+						
+						//console.info(`No data found, using channel infos: ${id} (${website})`);
+						
+						let streamInfo = {
+							id: id,
+							type: "channel",
+							contentId: contentId,
+							online: streamData.online,
+							website: website,
+							streamName: streamData.streamName,
+							streamStatus: streamData.streamStatus,
+							streamGame: streamData.streamGame,
+							streamOwnerLogo: streamData.streamOwnerLogo,
+							streamCategoryLogo: streamData.streamCategoryLogo,
+							streamCurrentViewers: streamData.streamCurrentViewers,
+							streamUrl: getStreamURL(website, id, contentId, true),
+							facebookID: streamData.facebookID,
+							twitterID: streamData.twitterID
+						}
+						streamInfo.streamSettings = streamList[id];
+						
+						panel.port.emit("updateData", streamInfo);
 					}
-					streamInfo.streamSettings = streamList[id];
-					
-					panel.port.emit("updateData", streamInfo);
 				} else {
 					for(let contentId in liveStatus[website][id]){
 						let streamData = liveStatus[website][id][contentId];
@@ -749,8 +754,7 @@ function updatePanelData(){
 		}
 	}
 	let notCheckedYet = false;
-	for(let i in websites){
-		let website = websites[i];
+	for(let website in websites){
 		var streamList = (new streamListFromSetting(website)).objData;
 		for(let id in streamList){
 			if(!(id in liveStatus[website])){
@@ -819,8 +823,6 @@ function handleChange(state) {
 }
 
 //Affichage des notifications en ligne / hors ligne
-var notifications = require("sdk/notifications");
-
 function copyLivestreamerCmd(data){
 	let cmd = `livestreamer ${getStreamURL(data.website, data.id, data.contentId, false)} ${getPreference("livestreamer_cmd_quality")}`;
 	let clipboard_success = clipboard.set(cmd, "text");
@@ -1077,8 +1079,7 @@ function doStreamNotif(website, id, contentId, streamSetting){
 function getOfflineCount(){
 	var offlineCount = 0;
 	
-	for(let i in websites){
-		let website = websites[i];
+	for(let website in websites){
 		let streamList = (new streamListFromSetting(website)).objData;
 		
 		for(let id in streamList){
@@ -1142,137 +1143,38 @@ function setIcon() {
 	}
 };
 
-let website_channel_id = /channel\:\:(.*)/;
-let facebookID_from_url = /(?:http|https):\/\/(?:www\.)?facebook.com\/([^\/]+)(?:\/.*)?/;
-let twitterID_from_url = /(?:http|https):\/\/(?:www\.)?twitter.com\/([^\/]+)(?:\/.*)?/;
-function API(website, id){
-	this.id = id;
-	this.url = "";
-	this.overrideMimeType = "";
-	
-	switch(website){
-		case "dailymotion":
-			if(website_channel_id.test(id)){
-				this.url = `https://api.dailymotion.com/videos?live_onair&owners=${website_channel_id.exec(id)[1]}&fields=id,title,owner,audience,url,mode,onair?_= ${new Date().getTime()}`;
-			} else {
-				this.url = `https://api.dailymotion.com/video/${id}?fields=title,owner,user.username,audience,url,mode,onair?_=${new Date().getTime()}`;
-			}
-			this.overrideMimeType = "text/plain; charset=latin1";
-			break;
-		case "hitbox":
-			this.url = `https://api.hitbox.tv/media/live/${id}`;
-			this.overrideMimeType = "text/plain; charset=utf-8";
-			break;
-		case "twitch":
-			this.url = `https://api.twitch.tv/kraken/streams/${id}`;
-			this.overrideMimeType = "application/vnd.twitchtv.v3+json; charset=utf-8"; //"text/plain; charset=utf-8";
-			break;
-		case "beam":
-			this.url = `https://beam.pro/api/v1/channels/${id}`;
-			this.overrideMimeType = "text/plain; charset=utf-8";
-			break;
-	}
-}
-function API_channelInfos(website, id){
-	this.id = id;
-	this.url = "";
-	this.overrideMimeType = "";
-	
-	switch(website){
-		case "dailymotion":
-			this.url = `https://api.dailymotion.com/user/${website_channel_id.exec(id)[1]}?fields=id,username,screenname,url,avatar_720_url,facebook_url,twitter_url`;
-			this.overrideMimeType = "text/plain; charset=latin1";
-			break;
-		default:
-			this.url = null;
-			this.overrideMimeType = null;
-	}
-}
-function API_second(website, id){
-	this.id = id;
-	this.url = "";
-	this.overrideMimeType = "";
-	
-	switch(website){
-		case "dailymotion":
-			this.url = `https://api.dailymotion.com/video/${id}?fields=id,user.screenname,game.title,user.avatar_720_url,user.facebook_url,user.twitter_url`;
-			this.overrideMimeType = "text/plain; charset=latin1";
-			break;
-		case "twitch":
-			this.url = `https://api.twitch.tv/kraken/users/${id}`;
-			this.overrideMimeType = "application/vnd.twitchtv.v3+json; charset=utf-8"; //"text/plain; charset=utf-8";
-			break;
-		default:
-			this.url = null;
-			this.overrideMimeType = null;
-	}
-}
-function importAPI(website, id){
-	this.id = id;
-	this.url = "";
-	this.overrideMimeType = "";
-	
-	switch(website){
-		case "dailymotion":
-			this.url = `https://api.dailymotion.com/user/${id}/following?fields=id,username,facebook_url,twitter_url?_=${new Date().getTime()}`;
-			this.overrideMimeType = "text/plain; charset=latin1";
-			break;
-		case "hitbox":
-			this.url = `https://api.hitbox.tv/following/user?user_name=${id}`;
-			this.overrideMimeType = "text/plain; charset=utf-8";
-			break;
-		case "twitch":
-			this.url = `https://api.twitch.tv/kraken/users/${id}/follows/channels`;
-			this.overrideMimeType = "application/vnd.twitchtv.v3+json; charset=utf-8";
-			break;
-		case "beam":
-			this.url = `https://beam.pro/api/v1/users/${id}/follows?limit=-1&fields=id,token`;
-			this.overrideMimeType = "text/plain; charset=utf-8";
-			break;
-	}
-}
+const website_channel_id = /channel\:\:(.*)/,
+	facebookID_from_url = /(?:http|https):\/\/(?:www\.)?facebook.com\/([^\/]+)(?:\/.*)?/,
+	twitterID_from_url = /(?:http|https):\/\/(?:www\.)?twitter.com\/([^\/]+)(?:\/.*)?/;
+
 function isValidResponse(website, data){
 	if(data == null || typeof data != "object" || JSON.stringify == "{}"){
 		console.warn("Unable to get stream state (no connection).");
 		return false;
 	}
-	switch(website){
-		case "dailymotion":
-			if(typeof data.error == "object"){
-				console.warn(`[${website}] Unable to get stream state (error detected).`);
-				return false;
-			} else if(typeof data.id == "string"){
-				return true;
-			} else if(data.mode != "live" && typeof data.list == "undefined"){
-				console.warn(`[${website}] Unable to get stream state (not a stream).`);
-				return false;
-			}
+	let state = websites[website].isValidResponse(data);
+	switch(state){
+		case "error":
+			console.warn(`[${website}] Unable to get stream state (error detected).`);
+			return false;
 			break;
-		case "hitbox":
-			if(data.error == "live"){
-				console.warn(`[${website}] Unable to get stream state (error detected).`);
-				return false;
-			}
-			if(data.error == true){
-				console.warn(`[${website}] Unable to get stream state (error detected).`);
-				return false;
-			}
+		case "notstream":
+			console.warn(`[${website}] Unable to get stream state (not a stream).`);
+			return false;
 			break;
-		case "twitch":
-			if(data.error == "Not Found"){
-				console.warn(`[${website}] Unable to get stream state (error detected).`);
-				return false;
-			}
-			break;
-		case "beam":
-			if(data == "Channel not found." || data.statusCode == 404){
-				console.warn(`[${website}] Unable to get stream state (error detected).`);
-				return false;
-			}
+		case "":
+		case "success":
+			return true
+		default:
+			console.warn(`[${website}] Unknown validation state (${state}).`);
+			return true;
 			break;
 	}
+	
 	return true;
 }
+
+/*
 let checkingLives_processState = {data: {},checked: null, totalwebsites: null, total: null};
 function checkLiveEnd(){
 	let checked = checkingLives_processState.checked;
@@ -1290,34 +1192,30 @@ function checkLiveEnd(){
 			delete data[website];
 		}
 	}
-	//if(totalwebsites == websites.length && checked == total){
 	if(totalwebsites == websites.length && JSON.stringify(data) == "{}"){
-		for(let i in websites){
-			let website = websites[i];
+		for(let website in websites){
 			var streamList = (new streamListFromSetting(website)).objData;
 			for(let id in liveStatus[website]){
 				if(id in streamList && (typeof streamList[id].ignore == "boolean" && streamList[id].ignore == true)){
-					console.dir(liveStatus[website][id]);
 					console.info("[Live notifier] Live check end");
 					//doStreamNotif(website, id, streamList[id].contentId, streamList[id].streamSetting);
 				}
 			}
 		}
 	}
-}
+}*/
 function checkLives(){
 	console.group();
 	
-	checkingLives_processState.data = {};
-	checkingLives_processState.checked = 0;
-	checkingLives_processState.total = 0;
+	//checkingLives_processState.data = {};
+	//checkingLives_processState.checked = 0;
+	//checkingLives_processState.total = 0;
 	
-	for(let i in websites){
-		let website = websites[i];
+	for(let website in websites){
 		let streamList = (new streamListFromSetting(website)).objData;
 		
-		checkingLives_processState.totalwebsites += streamList.totalwebsites + 1;
-		checkingLives_processState.total += streamList.length;
+		//checkingLives_processState.totalwebsites += streamList.totalwebsites + 1;
+		//checkingLives_processState.total += streamList.length;
 		
 		console.group();
 		console.info(website);
@@ -1325,10 +1223,11 @@ function checkLives(){
 		console.groupEnd();
 		
 		for(let id in streamList){
-			if(typeof checkingLives_processState.data[website]){
+			/*if(typeof checkingLives_processState.data[website] == "undefined"){
 				checkingLives_processState.data[website] = {};
 			}
-			checkingLives_processState.data[website][id] = 0;
+			checkingLives_processState.data[website][id] = 0;*/
+			
 			if(typeof streamList[id].ignore == "boolean" && streamList[id].ignore == true){
 				//console.info(`Ignoring ${id}`);
 				continue;
@@ -1345,102 +1244,99 @@ function checkLives(){
 
 
 function getPrimary(id, website, streamSetting, url, pageNumber){
-	let current_API = new API(website, id);
+	let current_API = websites[website].API(id);
 	if(typeof url == "string"){
 		current_API.url = url;
 	}
 	
-	console.time(id);
+	console.time(`${website}::${id}`);
 	
-	Request({
+	Request_Get({
 		url: current_API.url,
 		overrideMimeType: current_API.overrideMimeType,
 		onComplete: function (response) {
-			let id = current_API.id;
-			data = response.json;
+			let data = response.json;
 			
 			console.group();
-			console.info(`${website} - ${id} (${current_API.url})`);
+			console.info(`${website} - ${id} (${response.url})`);
 			console.dir(data);
 			
 			if(typeof liveStatus[website][id] == "undefined"){
 				liveStatus[website][id] = {};
 			}
 			
-			if(website_channel_id.test(id)){
-				if(typeof pageNumber == "number"){
-					pagingPrimary[website](id, website, streamSetting, data, pageNumber)
+			if(website_channel_id.test(id) == true){
+				if(isValidResponse(website, data) == true){
+					let streamListData;
+					if(typeof pageNumber == "number"){
+						streamListData = websites[website].channelList(id, website, data, pageNumber);
+					} else {
+						streamListData = websites[website].channelList(id, website, data);
+					}
+					
+					if(JSON.stringify(streamListData.streamList) == "{}"){
+						getChannelInfo(website, id);
+						channelListEnd(website, id);
+					} else {
+						for(let i in streamListData.streamList){
+							let contentId = i;
+							//checkingLives_processState.data[website][id] += 1;
+							processPrimary(id, contentId, website, streamSetting, streamListData.streamList[i]);
+						}
+						if(streamListData.hasOwnProperty("next") == true){
+							if(streamListData.next == null){
+								channelListEnd(website, id);
+							} else {
+								getPrimary(id, website, streamSetting, streamListData.url, streamListData.next_page_number);
+							}
+						}
+					}
 				} else {
-					pagingPrimary[website](id, website, streamSetting, data)
+					channelListEnd(website, id);
 				}
 			} else {
-				checkingLives_processState.data[website][id] = 1;
 				processPrimary(id, id, website, streamSetting, data);
+				//checkingLives_processState.data[website][id] = 1;
 			}
 		}
-	}).get();
+	});
 }
 
-function pagingPrimaryEnd(id){
+function channelListEnd(website, id){
 	setIcon();
-	console.timeEnd(id);
+	console.timeEnd(`${website}::${id}`);
 	console.groupEnd();
-}
-let pagingPrimary = {
-	"dailymotion":
-		function(id, website, streamSetting, data, pageNumber){
-			let list = data.list;
-			
-			if(data.total == 0){
-				getChannelInfo(website, id);
-				pagingPrimaryEnd(id);
-			} else {
-				for(let i in list){
-					let contentId = list[i].id;
-					checkingLives_processState.data[website][id] += 1;
-					processPrimary(id, contentId, website, streamSetting, list[i]);
-				}
-				
-				if(data.has_more){
-					let next_url = (new API(website, website_channel_id.exec(id)[1])).url;
-					let next_page_number = ((typeof pageNumber == "number")? pageNumber : 1) + 1;
-					getPrimary(id, website, streamSetting, next_url + "&page=" + next_page_number, next_page_number);
-				} else {
-					pagingPrimaryEnd(id);
-				}
-			}
-		}
 }
 
 function processPrimary(id, contentId, website, streamSetting, data){
 	if(typeof liveStatus[website][id][contentId] == "undefined"){
-		liveStatus[website][id][contentId] = {"online": false, "notificationStatus": false, "streamName": contentId, "streamStatus": "", "streamGame": "", "streamOwnerLogo": "", "streamCategoryLogo": "", "streamCurrentViewers": null, "streamURL": "", "facebookID": "", "twitterID": ""};
+		let defaultStatus = liveStatus[website][id][contentId] = {"online": false, "notificationStatus": false, "streamName": contentId, "streamStatus": "", "streamGame": "", "streamOwnerLogo": "", "streamCategoryLogo": "", "streamCurrentViewers": null, "streamURL": "", "facebookID": "", "twitterID": ""};
 	}
 	if(isValidResponse(website, data) == true){
-		let liveState = checkLiveStatus[website](id, contentId, data);
+		let liveState = websites[website].checkLiveStatus(id, contentId, data, liveStatus[website][id][contentId]);
 		if(liveState !== null){
-			let second_API = new API_second(website, contentId);
-			
-			if(second_API.url !== null && second_API.overrideMimeType !== null){
-				Request({
+			if(websites[website].hasOwnProperty("API_second") == true){
+				let second_API = websites[website].API_second(contentId);
+				
+				Request_Get({
 					url: second_API.url,
 					overrideMimeType: second_API.overrideMimeType,
 					onComplete: function (response) {
 						let data_second = response.json;
 						
-						console.info(website + " - " + id + " (" + second_API.url + ")");
+						console.info(`${website} - ${id} (${response.url})`);
 						console.dir(data_second);
 						
-						seconderyInfo[website](id, contentId, data_second, liveState);
+						websites[website].seconderyInfo(id, contentId, data_second, liveStatus[website][id][contentId], liveState);
 						
 						doStreamNotif(website, id, contentId, streamSetting, liveState);
-						checkLiveEnd();
+						//checkLiveEnd();
 						setIcon();
 					}
-				}).get();
+				});
 			} else {
 				doStreamNotif(website, id, contentId, streamSetting, liveState);
-				checkLiveEnd();
+				//checkLiveEnd();
 			}
 		} else {
 			console.warn("Unable to get stream state.");
@@ -1448,208 +1344,36 @@ function processPrimary(id, contentId, website, streamSetting, data){
 	}
 	setIcon();
 	
-	console.timeEnd(id);
+	console.timeEnd(`${website}::${id}`);
 	console.groupEnd();
 }
 function getChannelInfo(website, id){
-	let channelInfos_API = new API_channelInfos(website, id);
+	let channelInfos_API = websites[website].API_channelInfos(id);
 	
-	if(typeof channelInfos["dailymotion"][id] == "undefined"){
-		channelInfos["dailymotion"][id] = {"online": false, "notificationStatus": false, "streamName": (website_channel_id.test(id) == true)? website_channel_id.exec(id)[1] : id, "streamStatus": "", "streamGame": "", "streamOwnerLogo": "", "streamCategoryLogo": "", "streamCurrentViewers": null, "streamURL": "", "facebookID": "", "twitterID": ""};
+	if(typeof channelInfos[website][id] == "undefined"){
+		let defaultChannelInfos = channelInfos["dailymotion"][id] = {"online": false, "notificationStatus": false, "streamName": (website_channel_id.test(id) == true)? website_channel_id.exec(id)[1] : id, "streamStatus": "", "streamGame": "", "streamOwnerLogo": "", "streamCategoryLogo": "", "streamCurrentViewers": null, "streamURL": "", "facebookID": "", "twitterID": ""};
 	}
-	
-	if(channelInfos_API.url !== null && channelInfos_API.overrideMimeType !== null){
-		Request({
+	if(websites[website].hasOwnProperty("API_second") == true){
+		Request_Get({
 			url: channelInfos_API.url,
 			overrideMimeType: channelInfos_API.overrideMimeType,
 			onComplete: function (response) {
 				let data_channelInfos = response.json;
 				
-				console.info(website + " - " + id + " (" + channelInfos_API.url + ")");
+				console.info(`${website} - ${id} (${response.url})`);
 				console.dir(data_channelInfos);
 				
 				if(isValidResponse(website, data_channelInfos) == true){
-					channelInfosProcess[website](id, data_channelInfos);
+					websites[website].channelInfosProcess(id, data_channelInfos, channelInfos[website][id]);
 				}
 			}
-		}).get();
+		});
 	}
-}
-let channelInfosProcess = {
-	"dailymotion":
-		function(id, data){
-			let streamData = channelInfos["dailymotion"][id];
-			if(data.hasOwnProperty("screenname")){
-				streamData.streamName = data["screenname"];
-				streamData.streamURL = data.url;
-				if(typeof data["avatar_720_url"] == "string" && data["avatar_720_url"] != ""){
-					streamData.streamOwnerLogo = data["avatar_720_url"];
-				}
-				
-				if(typeof data["facebook_url"] == "string" && data["facebook_url"] != "" && facebookID_from_url.test(data["facebook_url"])){
-					streamData.facebookID = facebookID_from_url.exec(data["facebook_url"])[1];
-				}
-				if(typeof data["twitter_url"] == "string" && data["twitter_url"] != "" && twitterID_from_url.test(data["twitter_url"])){
-					streamData.facebookID = twitterID_from_url.exec(data["twitter_url"])[1];
-				}
-			}
-		}
-}
-
-//Fonction principale : check si le live est on
-let checkLiveStatus = {
-	"beam":
-		function(id, contentId, data){
-			let streamData = liveStatus["beam"][id][contentId];
-			
-			streamData.streamName = data.user["username"];
-			streamData.streamStatus = data["name"];
-			
-			if(typeof data.user["avatarUrl"] == "string" && data.user["avatarUrl"] != ""){
-				streamData.streamOwnerLogo = data.user["avatarUrl"];
-			}
-			streamData.streamCurrentViewers = parseInt(data["viewersCurrent"]);
-			if(typeof data.user.social["twitter"] == "string" && data.user.social["twitter"] != "" && twitterID_from_url.test(data.user.social["twitter"])){
-				streamData.twitterID = twitterID_from_url.exec(data.user.social["twitter"])[1];
-			}
-			
-			streamData.online = data["online"];
-			return streamData.online;
-		},
-	"dailymotion":
-		function(id, contentId, data){
-			let streamData = liveStatus["dailymotion"][id][contentId];
-			streamData.streamName = data.title;
-			streamData.streamCurrentViewers = JSON.parse(data.audience);
-			streamData.streamURL = data.url;
-			if(typeof data.onair == "boolean"){
-				streamData.online = data.onair;
-				return streamData.online;
-			} else {
-				return null;
-			}
-		},
-	"hitbox":
-		function(id, contentId, data){
-			let streamData = liveStatus["hitbox"][id][contentId];
-			if(data.hasOwnProperty("livestream") == false){
-				if(data.error_msg="no_media_found"){
-					streamData.online = false;
-				}
-				streamData.streamName = id;
-				return null;
-			}
-			if(typeof data["livestream"][0] == "object"){
-				data = data["livestream"][0];
-				streamData.streamName = data["media_user_name"];
-				streamData.streamStatus = data["media_status"];
-				if(typeof data["category_name"] == "string" && data["category_name"] != ""){
-					streamData.streamGame = data["category_name"];
-				}
-				if(data["category_logo_large"] !== null){
-					streamData.streamCategoryLogo = "http://edge.sf.hitbox.tv" + data["category_logo_large"];
-				} else if(data["category_logo_small"] !== null){
-					streamData.streamCategoryLogo = "http://edge.sf.hitbox.tv" + data["category_logo_small"];
-				} else {
-					streamData.streamCategoryLogo = "";
-				}
-				if(streamData.streamCategoryLogo = "http://edge.sf.hitbox.tv/static/img/generic/blank.gif"){
-					streamData.streamCategoryLogo = "";
-				}
-				if(typeof data.channel["user_logo"] !== "string" && data.channel["user_logo"].indexOf("/static/img/generic/default-user-") == -1){
-					streamData.streamOwnerLogo = "http://edge.sf.hitbox.tv" + data.channel["user_logo"];
-				} else if(typeof data.channel["user_logo"] !== "string" && data.channel["user_logo"].indexOf("/static/img/generic/default-user-") == -1){
-					streamData.streamOwnerLogo = "http://edge.sf.hitbox.tv" + data.channel["user_logo_small"];
-				} else {
-					streamData.streamOwnerLogo = "";
-				}
-				if(typeof data.channel["channel_link"] == "string" && data.channel["channel_link"] != ""){
-					streamData.streamURL = data.channel["channel_link"];
-				}
-				streamData.streamCurrentViewers = parseInt(data["media_views"]);
-				
-				streamData.online = (data["media_is_live"] == "1")? true : false;
-				if(typeof data.channel["twitter_account"] == "string" && data.channel["twitter_account"] != "" && typeof data.channel["twitter_account"] == "string" && data.channel["twitter_enabled"] == "1"){
-					streamData.twitterID = data.channel["twitter_account"];
-				}
-				return streamData.online;
-				/* if(data["media_is_live"] == "1"){
-					return true;
-				} else {
-					return false
-				}*/
-			} else {
-				return null;
-			}
-		},
-	"twitch":
-		function(id, contentId, data){
-			let streamData = liveStatus["twitch"][id][contentId];
-			if(data.hasOwnProperty("stream")){
-				data = data["stream"];
-				if(data != null){
-					streamData.streamName = data["channel"]["display_name"];
-					streamData.streamStatus = (data["channel"]["status"] != null)? data["channel"]["status"] : "";
-					streamData.streamGame = (data["game"] !== null && typeof data["game"] == "string")? data["game"] : "";
-					if(typeof data.channel["logo"] == "string" && data.channel["logo"] != "") {
-						streamData.streamOwnerLogo = data.channel["logo"];
-					}
-					if(typeof data.channel["url"] == "string" && data.channel["url"] != "") {
-						streamData.streamURL = data.channel["url"];
-					}
-					streamData.streamCurrentViewers = parseInt(data["viewers"]);
-					
-					streamData.online = true;
-					return streamData.online;
-				} else {
-					if(streamData.streamName == ""){
-						streamData.streamName = id;
-					}
-					streamData.online = false;
-					return streamData.online;
-				}
-			} else {
-				return null;
-			}
-		}
-}
-seconderyInfo = {
-	"dailymotion":
-		function(id, contentId, data, isStreamOnline){
-			let streamData = liveStatus["dailymotion"][id][contentId];
-			if(data.hasOwnProperty("user.screenname")){
-				if(isStreamOnline){
-					streamData.streamStatus = streamData.streamName;
-					streamData.streamGame = (data["game.title"] !== null && typeof data["game.title"] == "string")? data["game.title"] : "";
-				}
-				if(typeof data["user.avatar_720_url"] == "string" && data["user.avatar_720_url"] != ""){
-					streamData.streamOwnerLogo = data["user.avatar_720_url"];
-				}
-				streamData.streamName = data["user.screenname"];
-				
-				if(typeof data["user.facebook_url"] == "string" && data["user.facebook_url"] != "" && facebookID_from_url.test(data["user.facebook_url"])){
-					streamData.facebookID = facebookID_from_url.exec(data["user.facebook_url"])[1];
-				}
-				if(typeof data["user.twitter_url"] == "string" && data["user.twitter_url"] != "" && twitterID_from_url.test(data["user.twitter_url"])){
-					streamData.facebookID = twitterID_from_url.exec(data["user.twitter_url"])[1];
-				}
-			}
-		},
-	"twitch":
-		function(id, contentId, data, isStreamOnline){
-			let streamData = liveStatus["twitch"][id][contentId];
-			if(typeof data["display_name"] == "string"){
-				streamData.streamName = data["display_name"];
-			}
-			if(typeof data["logo"] == "string" && data["logo"] != ""){
-				streamData.streamOwnerLogo = data["logo"];
-			}
-		}
 }
 
 function importButton(website){
 	if(website == "beam"){
-		Request({
+		Request_Get({
 			url: `https://beam.pro/api/v1/channels/${getPreference(`${website}_user_id`)}`,
 			overrideMimeType: "text/plain; charset=utf-8",
 			onComplete: function (response) {
@@ -1670,114 +1394,66 @@ function importButton(website){
 					importStreams(website, numerical_id);
 				}
 			}
-		}).get();
+		});
 	} else {
 		importStreams(website, getPreference(`${website}_user_id`));
 	}
 }
 function importStreams(website, id, url, pageNumber){
-	let current_API = new importAPI(website, id);
+	let current_API = websites[website].importAPI(id);
 	if(typeof url == "string" && url != ""){
 		current_API.url = url;
 	} else {
-		console.time(current_API.id);
+		console.time(`${website}::${id}`);
 	}
-	Request({
+	Request_Get({
 		url: current_API.url,
 		overrideMimeType: current_API.overrideMimeType,
 		onComplete: function (response) {
-			let id = current_API.id;
 			let data = response.json;
 			
 			console.group();
 			console.info(`${website} - ${id} (${current_API.url})`);
 			console.dir(data);
 			
+			let streamListSetting = new streamListFromSetting(website);
+			let streamList = streamListSetting.objData;
+			
+			let importStreamList_Data;
 			if(typeof pageNumber == "number"){
-				importStreamWebsites[website](id, data, pageNumber);
+				importStreamList_Data = websites[website].importStreamWebsites(id, data, streamListSetting, pageNumber);
 			} else {
-				importStreamWebsites[website](id, data);
+				importStreamList_Data = websites[website].importStreamWebsites(id, data, streamListSetting);
 			}
+			
+			
+			for(let id of importStreamList_Data.list){
+				streamListSetting.addStream(website, id, "");
+			}
+			streamListSetting.update();
+			
+			if(importStreamList_Data.hasOwnProperty("next") == true && importStreamList_Data.next != null){
+				if(importStreamList_Data.next.hasOwnProperty("pageNumber") == true){
+					importStreams(website, id, importStreamList_Data.next.url, importStreamList_Data.next.pageNumber);
+				} else {
+					importStreams(website, id, importStreamList_Data.next.url);
+				}
+			} else {
+				importStreamsEnd(website, id);
+			}
+			
 			console.groupEnd();
 			setTimeout(function(){
 				refreshPanel(false);
 			}, 5000);
 		}
-	}).get();
+	});
 }
-function importStreamsEnd(id){
+function importStreamsEnd(website, id){
 	setIcon();
-	console.timeEnd(id);
+	console.timeEnd(`${website}::${id}`);
 }
 
-let importStreamWebsites = {
-	"beam": function(id, data){
-		let streamListSetting = new streamListFromSetting("beam");
-		let streamList = streamListSetting.objData;
-		if(typeof data == "object"){
-			for(let item of data){
-				streamListSetting.addStream("beam", item["token"], "");
-			}
-			streamListSetting.update();
-		}
-		importStreamsEnd(id);
-	},
-	"dailymotion": function(id, data, pageNumber){
-		let streamListSetting = new streamListFromSetting("dailymotion");
-		let streamList = streamListSetting.objData;
-		
-		if(data.total > 0){
-			for(let item of data.list){
-				if(!streamListSetting.streamExist("dailymotion", `channel::${item.id}`) && !streamListSetting.streamExist("dailymotion", `channel::${item.username}`)){
-					streamListSetting.addStream("dailymotion", `channel::${item.id}`, "");
-				} else {
-					console.log(`${item.username} already exist`);
-				}
-			}
-			streamListSetting.update();
-		}
-		
-		if(data.has_more){
-			let next_url = new importAPI("dailymotion", id).url;
-			let next_page_number = ((typeof pageNumber == "number")? pageNumber : 1) + 1;
-			importStreams("dailymotion", id, next_url + "&page=" + next_page_number, next_page_number);
-		} else {
-			importStreamsEnd(id);
-		}
-	},
-	"hitbox": function(id, data, pageNumber){
-		let streamListSetting = new streamListFromSetting("hitbox");
-		let streamList = streamListSetting.objData;
-		if(typeof data.following == "object"){
-			for(let item of data.following){
-				streamListSetting.addStream("hitbox", item["user_name"], "");
-			}
-			streamListSetting.update();
-			if(data.following.length > 0){
-				let next_url = new importAPI("hitbox", id).url;
-				let next_page_number = ((typeof pageNumber == "number")? pageNumber : 1) + 1;
-				importStreams("hitbox", id, next_url + "&offset=" + next_page_number, next_page_number);
-			} else {
-				importStreamsEnd(id);
-			}
-		}
-	},
-	"twitch": function(id, data){
-		let streamListSetting = new streamListFromSetting("twitch");
-		let streamList = streamListSetting.objData;
-		if(typeof data.follows == "object"){
-			for(let item of data.follows){
-				streamListSetting.addStream("twitch", item["channel"]["display_name"], "");
-			}
-			streamListSetting.update();
-			if(data.follows.length > 0 && typeof data._links.next == "string"){
-				importStreams("twitch", id, data._links.next);
-			} else {
-				importStreamsEnd(id);
-			}
-		}
-	}
-}
 function importDailymotionButton(){importButton("dailymotion");}
 sp.on("dailymotion_import", importDailymotionButton);
 function importTwitchButton(){importButton("twitch");}